requirements.txt
flask
flask_login
flask_bcrypt
flask_sqlalchemy
wtform
email-validator